Remove connection trait
