+++ redis/src/test/scala/zio/redis/RedisConnectionSpec.scala
@@ -6,7 +6,7 @@ import zio.{Chunk, ZIO}
 
 import java.nio.charset.StandardCharsets
 
-object RedisConnectionLiveSpec extends BaseSpec {
+object RedisConnectionSpec extends BaseSpec {
   override def spec: Spec[Environment, Any] =
     suite("Redis Connection Byte stream")(
       test("can write and read") {
@@ -17,5 +17,5 @@ object RedisConnectionLiveSpec extends BaseSpec {
           res    <- stream.read.runHead
         } yield assert(res)(isSome(equalTo('*'.toByte)))
       }
-    ).provideLayer(RedisConnectionLive.local.orDie)
+    ).provideLayer(RedisConnection.local)
 }
